+++ source/CommonJobs/CommonJobs.Application.EvalForm/Commands/GetEvaluatedEmployees.cs
@@ -1,4 +1,4 @@
-﻿using CommonJobs.Application.EvalForm.EmployeeSearching;
+﻿using CommonJobs.Application.EvalForm.Indexes;
 using CommonJobs.Infrastructure.RavenDb;
 using Raven.Client.Linq;
 using System;
