+++ source/CommonJobs/CommonJobs.Application.EvalForm/Commands/GetEvaluatorEmployeesCommand.cs
@@ -1,4 +1,4 @@
-﻿using CommonJobs.Application.EvalForm.EmployeeSearching;
+﻿using CommonJobs.Application.EvalForm.Indexes;
 using CommonJobs.Application.Evaluations;
 using CommonJobs.Domain.Evaluations;
 using CommonJobs.Infrastructure.RavenDb;
